refactor the RuntimeConfig and fix the health_check flag
