+++ rust/sglang-server/src/runtime.rs
@@ -12,13 +12,15 @@
 //! Keeping CPU-bound tokenize/detokenize off the async executor avoids stalling
 //! axum's worker threads.
 
-use std::net::SocketAddr;
 use std::sync::{Arc, Mutex};
 use std::thread::JoinHandle;
 
 pub mod channels;
+mod config;
 pub mod ring;
 pub mod threads;
+
+pub use config::{RuntimeConfig, RustServerServerArgs, ServerArgs};
 
 use crate::runtime::channels::{DetokMsg, Senders, TmEvent};
 use crate::runtime::ring::{
@@ -29,174 +31,6 @@ use crate::{api_server, detokenizer, tokenizer, tokenizer_manager};
 
 // Re-export so stages keep importing `crate::runtime::Runnable`.
 pub use threads::Runnable;
-
-#[derive(Clone, Debug)]
-pub struct RuntimeConfig {
-    pub http_addr: SocketAddr,
-    pub api_worker_num: usize,
-    pub tokenizer_worker_num: usize,
-    pub detokenizer_worker_num: usize,
-    pub ingress_ring_cap: usize,
-    pub egress_ring_cap: usize,
-    pub channel_cap: usize,
-    /// CPU core ids the pools pin to (e.g. this rank's NUMA-local cores minus
-    /// the scheduler's reserved launch cores). `None` → run unpinned.
-    pub cores: Option<Vec<usize>>,
-    /// Path to a `tokenizer.json` (or a model dir containing one, a tiktoken
-    /// model file, or an HF Hub repo id resolved via the local cache). `None`
-    /// requires `skip_tokenizer_init`; otherwise startup is a hard error.
-    pub tokenizer_path: Option<String>,
-    /// HF revision used only when `tokenizer_path` is a repo id. `None` → main.
-    pub revision: Option<String>,
-    /// Static server metadata (server_args + model_config) for config endpoints.
-    /// `Arc` so cloning the config (and, downstream, each `AppState`) is cheap;
-    /// `ServerArgs` itself is immutable after construction.
-    pub server_args: Arc<ServerArgs>,
-}
-
-impl Default for RuntimeConfig {
-    fn default() -> Self {
-        Self {
-            http_addr: "127.0.0.1:30000".parse().unwrap(),
-            api_worker_num: 2,
-            tokenizer_worker_num: 2,
-            detokenizer_worker_num: 2,
-            ingress_ring_cap: 8192,
-            egress_ring_cap: 8192,
-            channel_cap: 8192,
-            cores: None,
-            tokenizer_path: None,
-            revision: None,
-            server_args: Arc::new(
-                ServerArgs::from_json("{}").expect("empty server_args blob parses"),
-            ),
-        }
-    }
-}
-
-/// The scheduler's startup blob (`RustServer._build_server_args`) parsed once into
-/// typed fields: values are post-`__post_init__`, unknown keys (e.g. `api_key`) are dropped.
-#[derive(Debug, serde::Deserialize)]
-pub struct ServerArgs {
-    /// HF repo id / local dir of the model, reported by `/get_model_info`.
-    #[serde(default)]
-    pub model_path: String,
-    /// Model name reported by `/v1/models` and `/server_info`.
-    #[serde(default)]
-    pub served_model_name: String,
-    /// Tokenizer source (model dir / `tokenizer.json` / HF repo id). Empty only
-    /// in minimal standalone blobs — then boot requires `skip_tokenizer_init`.
-    #[serde(default)]
-    pub tokenizer_path: String,
-    /// HF revision, used only when `tokenizer_path` is a repo id. `None` → main.
-    #[serde(default)]
-    pub revision: Option<String>,
-    /// HTTP bind address (see [`Self::bind`]).
-    #[serde(default = "default_host")]
-    pub host: String,
-    #[serde(default = "default_port")]
-    pub port: u16,
-    /// Log levels driving the access log — uvicorn runs at
-    /// `log_level_http or log_level` (see [`Self::http_access_log_enabled`]).
-    #[serde(default = "default_log_level")]
-    pub log_level: String,
-    #[serde(default)]
-    pub log_level_http: Option<String>,
-    /// Pinned tokenizer threads / detok shards (Python asserts both ≥ 1).
-    #[serde(default = "default_worker_num")]
-    pub tokenizer_worker_num: usize,
-    #[serde(default = "default_worker_num")]
-    pub detokenizer_worker_num: usize,
-    /// Token-ids-in / token-ids-out mode: no tokenizer load, raw `output_ids`
-    /// frames (drives the `Skip` detok backend and the ingress branch).
-    #[serde(default)]
-    pub skip_tokenizer_init: bool,
-    /// Streamed `/generate` frames carry per-step deltas instead of cumulative
-    /// text. Matches the Python `TokenizerManager`.
-    #[serde(default)]
-    pub incremental_streaming_output: bool,
-    /// The resolved Python `ModelConfig`, attached to the blob at dump time.
-    #[serde(default)]
-    pub model_config: ModelConfig,
-    /// Launch-time stamps (not `server_args` fields): sglang package version
-    /// and the scheduler-derived KV token capacity, reported by `/server_info`.
-    #[serde(default)]
-    pub version: Option<String>,
-    #[serde(default)]
-    pub max_total_num_tokens: Option<u64>,
-}
-
-/// The slice of the resolved Python `ModelConfig` the rust server reads.
-#[derive(Debug, Default, serde::Deserialize)]
-pub struct ModelConfig {
-    /// Resolved context length (`max_model_len` in `/v1/models`); mandatory at
-    /// boot ([`ServerArgs::validate_mandatory`]).
-    #[serde(default)]
-    pub context_len: Option<u64>,
-    /// Bounds client-supplied token ids — ingress 400s out-of-vocab ids before
-    /// they crash the scheduler's embedding lookup. `None` → unvalidated.
-    #[serde(default)]
-    pub vocab_size: Option<u64>,
-}
-
-fn default_host() -> String {
-    "127.0.0.1".into()
-}
-fn default_port() -> u16 {
-    30000
-}
-fn default_log_level() -> String {
-    "info".into()
-}
-fn default_worker_num() -> usize {
-    1
-}
-
-impl ServerArgs {
-    /// Parse the blob; errors on malformed JSON or a wrongly-typed field.
-    pub fn from_json(s: &str) -> Result<Self, String> {
-        serde_json::from_str(s).map_err(|e| e.to_string())
-    }
-
-    /// Fail fast at startup if a field an endpoint depends on is missing.
-    pub fn validate_mandatory(&self) -> Result<(), String> {
-        if self.served_model_name.is_empty() {
-            return Err("no 'served_model_name' in server_args".into());
-        }
-        if self.model_config.context_len.is_none() {
-            return Err("no resolvable context length (model_config.context_len)".into());
-        }
-        Ok(())
-    }
-
-    /// Bind address `host:port`. `host` is expected to be an IP — the result is
-    /// parsed as a `SocketAddr`.
-    pub fn bind(&self) -> String {
-        format!("{}:{}", self.host, self.port)
-    }
-
-    /// Whether the HTTP access log is emitted, mirroring the Python server:
-    /// uvicorn runs at `log_level_http or log_level` and prints access lines
-    /// only at info/debug. `--log-level-http warning` turns them off.
-    pub fn http_access_log_enabled(&self) -> bool {
-        let level = self
-            .log_level_http
-            .as_deref()
-            .filter(|s| !s.is_empty())
-            .unwrap_or(&self.log_level);
-        matches!(
-            level.to_ascii_lowercase().as_str(),
-            "trace" | "debug" | "info"
-        )
-    }
-
-    /// Pinned API threads for the embedded HTTP api-server. Python `server_args`
-    /// has no such field — this is derived: enough to cover the widest pool.
-    pub fn api_worker_num(&self) -> usize {
-        4.max(self.tokenizer_worker_num)
-            .max(self.detokenizer_worker_num)
-    }
-}
 
 /// Live runtime. Held by the pyo3 bridge; the Python boundary reads `ingress`
 /// and `egress`. `request_shutdown` (also run on `Drop`) stops every stage.
@@ -260,8 +94,8 @@ impl Drop for Runtime {
 pub fn start(cfg: RuntimeConfig) -> Result<Runtime, String> {
     // Bind the API server port before spawning any thread, so an unavailable
     // port (EADDRINUSE) is a hard startup error.
-    let listener = std::net::TcpListener::bind(cfg.http_addr)
-        .map_err(|e| format!("bind {} failed: {e}", cfg.http_addr))?;
+    let listener = std::net::TcpListener::bind(cfg.rust_server_args.http_addr)
+        .map_err(|e| format!("bind {} failed: {e}", cfg.rust_server_args.http_addr))?;
     listener
         .set_nonblocking(true)
         .map_err(|e| format!("listener set_nonblocking failed: {e}"))?;
@@ -272,16 +106,19 @@ pub fn start(cfg: RuntimeConfig) -> Result<Runtime, String> {
 
     // --- rings (Rust ↔ Python) ---
     let (ingress_tx, ingress_rx): (IngressProducer, IngressConsumer) =
-        ingress_ring(cfg.ingress_ring_cap);
-    let (egress_tx, egress_rx): (EgressProducer, EgressConsumer) = egress_ring(cfg.egress_ring_cap);
+        ingress_ring(cfg.rust_server_args.ingress_ring_cap);
+    let (egress_tx, egress_rx): (EgressProducer, EgressConsumer) =
+        egress_ring(cfg.rust_server_args.egress_ring_cap);
 
     // --- inter-stage channels ---
-    let (tm_tx, tm_rx) = flume::bounded::<TmEvent>(cfg.channel_cap);
-    let (tok_tx, tok_rx) = flume::bounded::<crate::message::Request>(cfg.channel_cap);
-    let mut detok_tx = Vec::with_capacity(cfg.detokenizer_worker_num);
-    let mut detok_rx = Vec::with_capacity(cfg.detokenizer_worker_num);
-    for _ in 0..cfg.detokenizer_worker_num {
-        let (tx, rx) = flume::bounded::<DetokMsg>(cfg.channel_cap);
+    let (tm_tx, tm_rx) = flume::bounded::<TmEvent>(cfg.rust_server_args.channel_cap);
+    let (tok_tx, tok_rx) =
+        flume::bounded::<crate::message::Request>(cfg.rust_server_args.channel_cap);
+    let detokenizer_worker_num = cfg.server_args.detokenizer_worker_num;
+    let mut detok_tx = Vec::with_capacity(detokenizer_worker_num);
+    let mut detok_rx = Vec::with_capacity(detokenizer_worker_num);
+    for _ in 0..detokenizer_worker_num {
+        let (tx, rx) = flume::bounded::<DetokMsg>(cfg.rust_server_args.channel_cap);
         detok_tx.push(tx);
         detok_rx.push(rx);
     }
@@ -299,8 +136,11 @@ pub fn start(cfg: RuntimeConfig) -> Result<Runtime, String> {
     // The same instance is shared by the tokenizer pool (encode) and the detok
     // shards (decode); `None` only under `skip_tokenizer_init`.
     let dyn_tokenizer = tokenizer::load_tokenizer(
-        cfg.tokenizer_path.as_deref(),
-        cfg.revision.as_deref(),
+        // Empty only in minimal standalone blobs (the Python dump always
+        // resolves it); empty → no tokenizer, allowed only under
+        // `skip_tokenizer_init`.
+        (!cfg.server_args.tokenizer_path.is_empty()).then_some(&*cfg.server_args.tokenizer_path),
+        cfg.server_args.revision.as_deref(),
         skip_tokenizer_init,
     )?;
 
@@ -341,7 +181,7 @@ pub fn start(cfg: RuntimeConfig) -> Result<Runtime, String> {
         spawn_pool(
             "tokenizer",
             tok_cores,
-            cfg.tokenizer_worker_num,
+            cfg.server_args.tokenizer_worker_num,
             &mut threads,
             |_i| tokenizer::TokenizerWorker::new(tok_rx.clone(), tm_tx.clone(), tokenizer.clone()),
         );
@@ -407,7 +247,9 @@ pub fn start(cfg: RuntimeConfig) -> Result<Runtime, String> {
             .name("api-runtime".into())
             .spawn(move || {
                 let mut builder = tokio::runtime::Builder::new_multi_thread();
-                builder.worker_threads(cfg.api_worker_num).enable_all();
+                builder
+                    .worker_threads(cfg.rust_server_args.api_worker_num)
+                    .enable_all();
                 if let Some(cores) = api_cores {
                     let next = std::sync::atomic::AtomicUsize::new(0);
                     builder.on_thread_start(move || {
@@ -421,7 +263,7 @@ pub fn start(cfg: RuntimeConfig) -> Result<Runtime, String> {
                 rt.block_on(api_server::serve(
                     listener,
                     senders,
-                    cfg.channel_cap,
+                    cfg.rust_server_args.channel_cap,
                     cfg.server_args.clone(),
                     // Egress heartbeat watched by `/health_generate`.
                     api_activity,
@@ -457,12 +299,12 @@ mod tests {
         // `skip_tokenizer_init` → no tokenizer/detok model load; minimal boot.
         let server_args = ServerArgs::from_json(r#"{"skip_tokenizer_init": true}"#).unwrap();
         let cfg = RuntimeConfig {
-            http_addr: addr,
-            api_worker_num: 1,
-            tokenizer_worker_num: 1,
-            detokenizer_worker_num: 1,
+            rust_server_args: RustServerServerArgs {
+                http_addr: addr,
+                api_worker_num: 1,
+                ..Default::default()
+            },
             server_args: Arc::new(server_args),
-            ..Default::default()
         };
         // Bind is synchronous in `start`, so the port is already accepting.
         let rt = start(cfg).expect("start runtime");
@@ -497,12 +339,12 @@ mod tests {
 
         let server_args = ServerArgs::from_json(r#"{"skip_tokenizer_init": true}"#).unwrap();
         let cfg = RuntimeConfig {
-            http_addr: addr,
-            api_worker_num: 1,
-            tokenizer_worker_num: 1,
-            detokenizer_worker_num: 1,
+            rust_server_args: RustServerServerArgs {
+                http_addr: addr,
+                api_worker_num: 1,
+                ..Default::default()
+            },
             server_args: Arc::new(server_args),
-            ..Default::default()
         };
         let rt = start(cfg).expect("start runtime");
 
@@ -542,12 +384,12 @@ mod tests {
 
         let server_args = ServerArgs::from_json(r#"{"skip_tokenizer_init": true}"#).unwrap();
         let cfg = RuntimeConfig {
-            http_addr: addr,
-            api_worker_num: 1,
-            tokenizer_worker_num: 1,
-            detokenizer_worker_num: 1,
+            rust_server_args: RustServerServerArgs {
+                http_addr: addr,
+                api_worker_num: 1,
+                ..Default::default()
+            },
             server_args: Arc::new(server_args),
-            ..Default::default()
         };
         let rt = start(cfg).expect("start runtime");
 
@@ -599,12 +441,12 @@ mod tests {
 
         let server_args = ServerArgs::from_json(r#"{"skip_tokenizer_init": true}"#).unwrap();
         let cfg = RuntimeConfig {
-            http_addr: addr,
-            api_worker_num: 1,
-            tokenizer_worker_num: 1,
-            detokenizer_worker_num: 1,
+            rust_server_args: RustServerServerArgs {
+                http_addr: addr,
+                api_worker_num: 1,
+                ..Default::default()
+            },
             server_args: Arc::new(server_args),
-            ..Default::default()
         };
         let err = match start(cfg) {
             Ok(_) => panic!("bind conflict must fail startup, got Ok"),
